pyproject.toml
[tool.poetry]
name = "ineedempathy"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.61.1"
uvicorn = "^0.11.8"
sqlalchemy = "^1.3.19"
python-multipart = "^0.0.5"
pydantic = {extras = ["email"], version = "^1.6.1"}
python-jose = {extras = ["cryptography"], version = "^3.2.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.2"}
emails = "^0.6"
jinja2 = "^2.11.2"
fastapi-camelcase = "^1.0.2"

[tool.poetry.dev-dependencies]
mypy = "^0.782"
flake8 = "^3.8.3"
alembic = "^1.4.3"
sqlalchemy-stubs = {url = "https://github.com/dropbox/sqlalchemy-stubs/archive/master.zip"}
pytest = "^6.1.1"
mock = "^4.0.2"
pytest-cov = "^2.10.1"
requests = "^2.24.0"
pdbpp = "^0.10.2"
bpython = "^0.19"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
